+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Additional flag to ensure param exists before attempting to render iframe
+
 ## [0.1.0] - 2020-01-15
 
 ### Added